added: The function 'type_correction' was created to convert an int to a list if there was a comparison between an int and a list to be made.
The function 'compare_type' returns 'int' if both items being compared are type
'int'.

Wrote a recursive function 'compare_packets' to iterate through and
compare the values between each list, returning True if the packets are
in th correct or and False if not. Recursive implementation was chosen
so the inner lists would be unpacked and checked in the same manner as
the outer.

The 'main' funtion prints the pair number if they are in the correct order and
finally prints the total of those previously printed numbers.